+++ src/Particular.ThroughputCollector/AuditThroughput/AuditThroughputCollectorHostedService.cs
@@ -1,9 +1,11 @@
 ﻿namespace Particular.ThroughputCollector.Audit
 {
+    using System.Text.Json.Nodes;
     using Contracts;
-    using Infrastructure;
     using Microsoft.Extensions.Hosting;
     using Microsoft.Extensions.Logging;
+    using Particular.ThroughputCollector.Infrastructure;
+    using Particular.ThroughputCollector.Shared;
     using Persistence;
     using ServiceControl.Api;
 
@@ -47,23 +49,22 @@
 
             try
             {
-                //var httpFactory = await HttpAuth.CreateHttpClientFactory(throughputSettings.BrokerSettingValues[ServiceControlSettings.API], logger, configureNewClient: c => c.Timeout = TimeSpan.FromSeconds(30), cancellationToken: cancellationToken);
-                //var primary = new ServiceControlClient("ServiceControl", throughputSettings.BrokerSettingValues[ServiceControlSettings.API], httpFactory, logger);
-                //await primary.CheckEndpoint(content => content.Contains("\"known_endpoints_url\"") && content.Contains("\"endpoints_messages_url\""), cancellationToken); //TODO do we need this since we know the SC url?
-                //var knownEndpoints = await ServiceControlCommands.GetKnownEndpoints(primary, logger, cancellationToken);
-                var knownEndpoints = await ServiceControlCommands.GetKnownEndpoints(configurationApi, endpointsApi, auditCountApi, logger);
+                await VerifyAuditInstances();
+
+                var knownEndpoints = await GetKnownEndpoints();
 
                 if (!knownEndpoints.Any())
                 {
-                    throw new HaltException(HaltReason.InvalidEnvironment, "Successfully connected to ServiceControl API but no known endpoints could be found.");
+                    logger.LogWarning("Successfully connected to ServiceControl API but no known endpoints could be found.");
                 }
 
                 foreach (var endpoint in knownEndpoints)
                 {
                     if (!await ThroughputRecordedForYesterday(endpoint.Name, utcYesterday))
                     {
+                        var auditCounts = await GetAuditCountForEndpoint(endpoint.UrlName);
                         //for each endpoint record the audit count for the day we are currently doing as well as any others that are available
-                        await dataStore.RecordEndpointThroughput(SCEndpointToEndpoint(endpoint));
+                        await dataStore.RecordEndpointThroughput(SCEndpointToEndpoint(endpoint, auditCounts));
                     }
                 }
             }
@@ -80,7 +81,7 @@
             return endpoint?.DailyThroughput?.Any(a => a.DateUTC == utcDateTime) ?? false;
         }
 
-        Endpoint SCEndpointToEndpoint(ServiceControlEndpoint scEndpoint)
+        Endpoint SCEndpointToEndpoint(ServiceControlEndpoint scEndpoint, List<AuditCount> auditCounts)
         {
             return new Endpoint
             {
@@ -88,8 +89,104 @@
                 SanitizedName = EndpointNameSanitizer.SanitizeEndpointName(scEndpoint.Name, throughputSettings.Broker),
                 ThroughputSource = ThroughputSource.Audit,
                 EndpointIndicators = [EndpointIndicator.KnownEndpoint.ToString()],
-                DailyThroughput = scEndpoint.AuditCounts.Any() ? scEndpoint.AuditCounts.Select(c => new EndpointThroughput { DateUTC = c.UtcDate, TotalThroughput = c.Count }).ToList() : []
+                DailyThroughput = auditCounts.Any() ? auditCounts.Select(c => new EndpointThroughput { DateUTC = c.UtcDate, TotalThroughput = c.Count }).ToList() : []
             };
+        }
+
+        async Task<ServiceControlEndpoint[]> GetKnownEndpoints()
+        {
+            var endpoints = await endpointsApi.GetEndpoints();
+
+            var scEndpoints = endpoints?.Select(endpoint => new
+            {
+                Name = endpoint.Name ?? "",
+                HeartbeatsEnabled = endpoint.Monitored
+            })
+            .GroupBy(x => x.Name)
+            .Select(g => new ServiceControlEndpoint
+            {
+                Name = g.Key!,
+                HeartbeatsEnabled = g.Any(e => e.HeartbeatsEnabled),
+            })
+            .ToArray();
+
+            return scEndpoints ?? Array.Empty<ServiceControlEndpoint>();
+        }
+
+        async Task<List<AuditCount>> GetAuditCountForEndpoint(string endpointUrlName)
+        {
+            return (await auditCountApi.GetEndpointAuditCounts(page: null, pageSize: null, endpoint: endpointUrlName)).Select(s =>
+            {
+                return new AuditCount { Count = s.Count, UtcDate = s.UtcDate };
+            }).ToList();
+        }
+
+        async Task VerifyAuditInstances()
+        {
+            // Verify audit instances also have audit counts
+            var remotes = await configurationApi.GetRemoteConfigs();
+            var remotesInfo = new List<RemoteInstanceInformation>();
+            var valueType = remotes.GetType();
+
+            if (remotes != null && valueType.IsArray)
+            {
+                var remoteObjects = (object[])remotes;
+                if (remoteObjects.Length > 0)
+                {
+                    var props = remoteObjects[0].GetType().GetProperties();
+
+                    var apiUriProp = props.FirstOrDefault(w => w.Name == "ApiUri");
+                    var VersionProp = props.FirstOrDefault(w => w.Name == "Version");
+                    var statusProp = props.FirstOrDefault(w => w.Name == "Status");
+                    var configurationProp = props.FirstOrDefault(w => w.Name == "Configuration");
+
+                    foreach (var remote in remoteObjects)
+                    {
+                        var config = configurationProp != null ? configurationProp.GetValue(remote) as JsonNode : null;
+                        string? retention = null;
+                        if (config != null)
+                        {
+                            retention = config?.AsObject().TryGetPropertyValue("data_retention", out var dataRetention) == true &&
+                                        dataRetention?.AsObject().TryGetPropertyValue("audit_retention_period", out var auditRetentionPeriod) == true
+                                        ? auditRetentionPeriod!.GetValue<string>()
+                                        : null;
+                        }
+
+                        var remoteInstance = new RemoteInstanceInformation
+                        {
+                            ApiUri = apiUriProp != null ? apiUriProp.GetValue(remote)?.ToString() : "",
+                            VersionString = VersionProp != null ? VersionProp.GetValue(remote)?.ToString() : "",
+                            Status = statusProp != null ? statusProp.GetValue(remote)?.ToString() : "",
+                            Retention = TimeSpan.TryParse(retention, out var ts) ? ts : TimeSpan.Zero
+                        };
+
+                        remoteInstance.SemVer = SemVerVersion.TryParse(remoteInstance.VersionString, out var v) ? v : null;
+
+                        remotesInfo.Add(remoteInstance);
+                    }
+                }
+            }
+
+            foreach (var remote in remotesInfo)
+            {
+                if (remote.Status == "online" || remote.SemVer is not null)
+                {
+                    logger.LogInformation($"ServiceControl Audit instance at {remote.ApiUri} detected running version {remote.SemVer}");
+                }
+                else
+                {
+                    logger.LogWarning($"Unable to determine the version of one or more ServiceControl Audit instances. For the instance with URI {remote.ApiUri}, the status was '{remote.Status}' and the version string returned was '{remote.VersionString}'.");
+                    ;
+                }
+            }
+
+            // Want 2d audit retention so we get one complete UTC day no matter what time it is.
+            // Customers are expected to run at least version 4.29 for their Audit instances
+            var allHaveAuditCounts = remotesInfo.All(r => r.SemVer?.Version >= MinAuditCountsVersion && r.Retention >= TimeSpan.FromDays(2));
+            if (!allHaveAuditCounts)
+            {
+                logger.LogWarning($"At least one ServiceControl Audit instance is either not running the required version ({MinAuditCountsVersion}) or is not configured for at least 2 days of retention. Audit throughput will not be available.");
+            }
         }
 
         readonly ILogger logger;
@@ -98,5 +195,7 @@
         IConfigurationApi configurationApi;
         IAuditCountApi auditCountApi;
         IEndpointsApi endpointsApi;
+
+        static readonly Version MinAuditCountsVersion = new Version(4, 29);
     }
 }
