+++ src/Particular.ThroughputCollector/Contracts/ServiceControlEndpoint.cs
@@ -7,6 +7,5 @@
         public string Name { get; set; } = string.Empty;
         public string UrlName => WebUtility.UrlEncode(Name);
         public bool HeartbeatsEnabled { get; set; }
-        public AuditCount[] AuditCounts { get; set; } = Array.Empty<AuditCount>();
     }
 }
